Add files via upload
This code shows **Fibonacci series with and without recursion** 
OBSERVATION:
since every time the function is called , for 0 ,1,1 the series takes same time with no difference as it doesnt call any recurive func. after every iteration in the loop for n-2 & n-1 for a function call the time increases gradually and then mounts to the half U shape as the number of calls being made to the recursive function is something of 2^n, or exponential complexity. With a complexity of 2^n, and an n of 40, that would be 240, or 1,099,511,627,776, or about 1 trillion operations, for first 30 numbers of series 2^30 is 1,073,741,824, which would take only about 1 second.